+++ src/main/java/net/medievalweapons/mixin/client/BufferBuilderStorageMixin.java
@@ -1,0 +1,25 @@
+package net.medievalweapons.mixin.client;
+
+import it.unimi.dsi.fastutil.objects.Object2ObjectLinkedOpenHashMap;
+import net.fabricmc.api.EnvType;
+import net.fabricmc.api.Environment;
+import net.medievalweapons.init.RenderInit;
+import net.minecraft.client.render.BufferBuilderStorage;
+import net.minecraft.client.render.RenderLayer;
+import net.minecraft.client.util.BufferAllocator;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+@Environment(EnvType.CLIENT)
+@Mixin(BufferBuilderStorage.class)
+public class BufferBuilderStorageMixin {
+
+    @Inject(method = "assignBufferBuilder", at = @At("HEAD"))
+    private static void assignBufferBuilder(Object2ObjectLinkedOpenHashMap<RenderLayer, BufferAllocator> builderStorage, RenderLayer layer, CallbackInfo info) {
+        for (RenderLayer customLayer : RenderInit.CUSTOM_GLINT.values()) {
+            builderStorage.put(customLayer, new BufferAllocator(256));
+        }
+    }
+}